Create Accessibility - Career Launch 2024
"Rate My Path" is a platform designed for University of Maryland students with disability, allowing them to view others' ratings of various campus facilities and routes in order to help them navigate campus safely and efficiently. It aims to reduce the challenges and frustrations faced by disabled students when presented with inaccessible routes and facilities that they might otherwise be unaware of before experiencing them.